refactor(rebuild): extract gateway provider preflight
