+++ src/lib/actions/sandbox/rebuild.ts
@@ -22,9 +22,6 @@ const hermesProviderAuth = require("../../hermes-provider-auth") as {
     baseUrl?: string,
   ) => void;
 };
-const { providerExistsInGateway } = require("../../onboard/providers") as {
-  providerExistsInGateway: (name: string, runOpenshellFn: typeof runOpenshell) => boolean;
-};
 
 import {
   detectOpenShellStateRpcPreflightIssue,
@@ -80,6 +77,10 @@ import {
   resolveRebuildLiveState,
 } from "./rebuild-flow-helpers";
 import { buildRebuildRecreateOnboardOpts } from "./rebuild-gpu-opt-out";
+import {
+  checkRebuildGatewayProviderOrBail,
+  shouldVerifyRebuildGatewayProvider,
+} from "./rebuild-provider-preflight";
 import {
   getRebuildCredentialEnvFromRegistry,
   isLocalInferenceProvider,
@@ -210,52 +211,6 @@ function preflightHermesProviderCredentials(
   }
   console.error("");
   console.error("  Sandbox is untouched — no data was lost.");
-  return false;
-}
-
-function printMissingRebuildGatewayProvider(provider: string, credentialEnv: string | null): void {
-  console.error("");
-  console.error(
-    `  ${_RD}Rebuild preflight failed:${R} provider '${provider}' is not registered in OpenShell.`,
-  );
-  console.error("  The sandbox registry still points at this upstream provider,");
-  console.error("  so rebuild will not recreate it before destroying the sandbox.");
-  if (credentialEnv) {
-    console.error(`  Rebuild cannot rely on ${credentialEnv} while that provider is missing.`);
-  }
-  console.error("");
-  console.error("  Re-register the provider in OpenShell or rerun onboard, then retry rebuild.");
-  console.error("  Sandbox is untouched — no data was lost.");
-}
-
-function shouldVerifyRebuildGatewayProvider(
-  provider: string | null | undefined,
-): provider is string {
-  return Boolean(
-    provider &&
-      !isLocalInferenceProvider(provider) &&
-      provider !== hermesProviderAuth.HERMES_PROVIDER_NAME,
-  );
-}
-
-function checkRebuildGatewayProviderOrBail(
-  provider: string | null | undefined,
-  credentialEnv: string | null,
-  log: (msg: string) => void,
-  bail: (msg: string, code?: number) => never,
-): boolean {
-  if (!shouldVerifyRebuildGatewayProvider(provider)) return true;
-
-  const providerRegisteredInGateway = providerExistsInGateway(provider, runOpenshell);
-  log(
-    `Preflight gateway provider check: provider '${provider}' is ${
-      providerRegisteredInGateway ? "registered" : "missing"
-    } in OpenShell`,
-  );
-  if (providerRegisteredInGateway) return true;
-
-  printMissingRebuildGatewayProvider(provider, credentialEnv);
-  bail(`Missing gateway provider: ${provider}`);
   return false;
 }
